Refactor import paths from open_ai to openai for consistency
