+++ internal/application/service/assistant.go
@@ -3,7 +3,7 @@ package service
 import (
 	"log"
 	"nekoacm-server/internal/application/dto"
-	"nekoacm-server/internal/infrastructure/open_ai"
+	"nekoacm-server/internal/infrastructure/openai"
 	"nekoacm-server/prompt"
 	"strings"
 )
@@ -17,7 +17,7 @@ func AssistantChat(msg dto.ChatMsg) (string, error) {
 	log.Println("请求对话，内容长度:", len(msg.Content))
 
 	// 请求模型
-	resp, err := open_ai.Chat(prompt.ChatAssistant, msg.Content)
+	resp, err := openai.Chat(prompt.ChatAssistant, msg.Content)
 	if err != nil {
 		return "", err
 	}
