+++ internal/infrastructure/openai/openai.go
@@ -1,4 +1,4 @@
-package open_ai
+package openai
 
 import (
 	"github.com/sashabaranov/go-openai"
