+++ internal/infrastructure/openai/service.go
@@ -1,4 +1,4 @@
-package open_ai
+package openai
 
 import (
 	"github.com/sashabaranov/go-openai"
